Integrate http client to modules view
