+++ Software Management Course Website/Controllers/CoursesController.cs
@@ -1,31 +1,45 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Dynamic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Newtonsoft.Json;
+using Microsoft.Extensions.Configuration;
+using serviceclient;
+using serviceclient.types;
 
 namespace Software_Management_Course_Website.Controllers
 {
     public class CoursesController : Controller
     {
         private Models.RootObject items;
+        
+        private static string endpoint = "http://css566backend.azurewebsites.net/";
 
-        public CoursesController()
+        private ServiceClient<Course> client = null;
+    
+        public List<Course> Courses { get; private set; } 
+
+        public string Message { get; set; }
+
+        public CoursesController(IConfiguration configuration)
         {
-            // Load JSON here
-            using (StreamReader r = new StreamReader(@"Data/AgileGameDevelopmentModule.json"))
-            {
-                string json = r.ReadToEnd();
-                items = JsonConvert.DeserializeObject<Models.RootObject>(json);
-            }
+            // test for presence of keys
+            endpoint = configuration["backendapi:uri"];
+            Message = configuration["backendapi:message"];
+
+            client = new ServiceClient<Course>(endpoint + "/course");
         }
 
         public IActionResult Index()
         {
-            return View(items);
+            var model = new CoursesViewModel();
+
+            model.Courses = client.Get();
+            model.Message = string.IsNullOrEmpty(Message) ? "message not set" : Message;
+
+            return View(model);
         }
     }
 }
