+++ Software Management Course Website/Views/Modules/Index.cshtml
@@ -1,16 +1,91 @@
-@model ModulesViewModel
-
+﻿@model RootObject
 @{
-    ViewData["title"] = "Modules"; 
-    Layout = "_Layout";
+    ViewData["Title"] = "Course Content";
 }
 
-<div class="module">
-    <h3>@Model.Message</h3>
+@*<h2>Course Content</h2>*@
 
-    <ul>
-    @foreach (var module in Model.Modules) {
-        <li>@module.ModuleId @module.Title</li>
+@if (Model == null || Model.Module == null)
+{
+    <h2><strong>Sorry, no content is available at this time :(</strong></h2>
+}
+else
+{
+    <p>
+        <h2>@Model.Module.Title</h2><br />
+        <h3>Author: @Model.Module.Author</h3><br />
+        <h4>Courses in this Module:</h4>
+        @foreach (var course in Model.Module.CourseIds)
+        {
+            <h5>
+                @course
+            </h5>
+        }
+    </p>
+    @foreach (var layoutElement in Model.Module.Layout)
+    {
+        switch (layoutElement.Type)
+        {
+            case "image":
+                <br />
+                <div class="image">
+                    <img src=@layoutElement.Link
+                         style="width: 50%; height: 50%;"
+                         alt="Agile Game Development with Scrum" />
+                    <div>@layoutElement.Caption</div>
+                </div>
+                <br />
+                break;
+
+            case "text":
+                <br />
+                <p>
+                    <h4>@layoutElement.Heading</h4>
+                    @Html.Raw(layoutElement.Text)
+                </p>
+                <br />
+                break;
+
+            case "video":
+                <br />
+                <p>
+                    <h4>@layoutElement.Heading</h4>
+                    <br />
+                    <iframe width="560"
+                            height="315"
+                            src=@layoutElement.Link
+                            frameborder="0"
+                            allow="autoplay; encrypted-media"
+                            allowfullscreen></iframe>
+                </p>
+                <br />
+                break;
+
+            case "slideset":
+                <br />
+                <h4>@layoutElement.Heading</h4>
+                <br />
+                <iframe src=@layoutElement.Link
+                        width='722px' height='565px'
+                        frameborder='0'></iframe>
+                <br />
+                break;
+
+            case "quiz":
+                <br />
+                <h4>@layoutElement.Heading</h4>
+                <br />
+                break;
+
+            case "custom":
+                <br />
+                <h4>@layoutElement.Heading</h4>
+                @Html.Raw(layoutElement.Link)
+                <br />
+                break;
+
+            default:
+                break;
+        }
     }
-    </ul>
-</div>
+}
