+++ Software Management Course Website/appsettings.json
@@ -9,7 +9,7 @@
     }
   },
   "backendapi": {
-    "uri": "http://css566backend.azurewebsites.net/module",
+    "uri": "http://css566backend.azurewebsites.net/",
     "docs": "docsuri",
     "message": "Module courses:"
   }
